Added nonzero option for single-dimension collapses
Updated the various operations in collapse so that "nonzero" now works for any single-dimension operation. Zero values are ignored unless they span an entire "strip" in the selected dimension, in which case they are almost always treated like a row of zeros (the exception being product, which treats them as 1s).